Update: algorithm list 분리
