+++ src/components/main/main.module.scss
@@ -8,24 +8,3 @@
     align-items: center;
   }
 }
-
-.algorithms {
-  position: relative;
-  width: 100%;
-  margin-top: 130px;
-  padding-left: 20px;
-  margin-bottom: 100px;
-  @media (max-width: c.$laptop) {
-    padding: 0;
-    margin-top: 20px;
-    width: calc(100vw - 40px);
-  }
-}
-
-.heading {
-  margin-bottom: 20px;
-  font-weight: 700;
-  @media (max-width: c.$laptop) {
-    margin: 15px 0 30px 0;
-  }
-}
